Bootleg home module export with factory function
Need to make it so this home module, and the rest to come, dont
auto load and display on the document. Instead this home module,
and the rest, should return an element that the caller (index.js)
can then append it themselves to wherever they want to. Also not sure
if I should probably instead explicitily need to load each module with
.renderAll(). Since they wont be appended anyways. But yea its prob
better to explicitily load them.