Base Tab Switching on 2 Tabs working
Need to change the classes given to each module's div. Each should
have a different class, instead of all sharing a .main class. If I
then want them to share style, can easily do that with css selectors

They need to be separate because we might want to display modules in
different styles or sizes, and it would be impossible to style
each module specifically different, while keeping them on a
div with the same class. So just make a diff class for each.